requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
python-dotenv
requests
pydantic
boto3
python-socketio[asyncio]
sib-api-v3-sdk

